requirements-dev.txt
pytest==5.4.1
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.7.5
pytest-cov
pylama
codecov

sphinx>=2.0.0
sphinx_rtd_theme
sphinxcontrib-images

autoflake
autopep8
coverage
black
flake8
isort
bandit
